+++ llvm/lib/Transforms/Instrumentation/PGOInstrumentation.cpp
@@ -116,7 +116,6 @@
 #include <cstdint>
 #include <memory>
 #include <numeric>
-#include <stdint.h>
 #include <string>
 #include <unordered_map>
 #include <utility>
@@ -1831,7 +1830,7 @@ void llvm::setProfMetadata(Module *M, Instruction *TI,
   MDBuilder MDB(M->getContext());
   assert(MaxCount > 0 && "Bad max count");
   uint64_t Scale = calculateCountScale(MaxCount);
-  SmallVector<uint64_t, 4> Weights;
+  SmallVector<unsigned, 4> Weights;
   for (const auto &ECI : EdgeCounts)
     Weights.push_back(scaleBranchCount(ECI, Scale));
